refactor: move createStatementData() to its own file
